ref: improve user model
