+++ src/resolvers/cohort.resolvers.ts
@@ -6,7 +6,8 @@ import { checkUserLoggedIn } from '../helpers/user.helpers'
 import Cohort from '../models/cohort.model'
 import Program from '../models/program.model'
 import Phase from '../models/phase.model'
-import { Organization, User } from '../models/user'
+import { User } from '../models/user'
+import { Organization } from '../models/organization.model'
 import { Context } from './../context'
 import { ProgramType } from './program.resolvers'
 import { OrganizationType } from './userResolver'
