+++ src/resolvers/coordinatorResolvers.ts
@@ -2,7 +2,9 @@ import { ApolloError } from 'apollo-server'
 import Cohort from '../models/cohort.model'
 import * as jwt from 'jsonwebtoken'
 import { ValidationError } from 'apollo-server'
-import { Attendance, Organization, User } from '../models/user'
+import { User } from '../models/user'
+import { Attendance } from '../models/attendance.model'
+import { Organization } from '../models/organization.model'
 import { checkUserLoggedIn } from '../helpers/user.helpers'
 import { checkLoggedInOrganization } from '../helpers/organization.helper'
 import getOrganizationTemplate from '../utils/templates/getOrganizationTemplate'
