+++ src/resolvers/resolver.ts
@@ -1,7 +1,8 @@
 import bcrypt from 'bcryptjs';
 import * as jwt from 'jsonwebtoken';
 import mongoose from 'mongoose';
-import { Profile, User } from '../models/user';
+import { User } from '../models/user';
+import { Profile } from '../models/profile.model';
 
 const SECRET = process.env.SECRET || 'test_secret';
