+++ src/resolvers/team.resolvers.ts
@@ -5,7 +5,8 @@ import Team from '../models/team.model';
 import Program from '../models/program.model';
 import Phase from '../models/phase.model';
 import Cohort from '../models/cohort.model';
-import { Organization, User } from '../models/user';
+import { User } from '../models/user';
+import { Organization } from '../models/organization.model';
 import { Context } from '../context';
 import { ProgramType } from './program.resolvers';
 import { OrganizationType } from './userResolver';
