+++ src/resolvers/userResolver.ts
@@ -8,7 +8,9 @@ import { checkUserLoggedIn } from '../helpers/user.helpers'
 import { checkLoggedInOrganization } from '../helpers/organization.helper'
 import Cohort from '../models/cohort.model'
 import Program from '../models/program.model'
-import { Organization, Profile, User, UserRole } from '../models/user'
+import { User, UserRole } from '../models/user'
+import { Organization } from '../models/organization.model'
+import { Profile } from '../models/profile.model'
 import { sendEmail } from '../utils/sendEmail'
 import organizationCreatedTemplate from '../utils/templates/organizationCreatedTemplate'
 import organizationApprovedTemplate from '../utils/templates/organizationApprovedTemplate'
